third_party/WebKit/LayoutTests/fast/dom/shadow/shadowdom-for-form-associated-element-useragent.html

<!DOCTYPE html>
<html>
<body>
<script src="../../../resources/js-test.js"></script>

<p>When a OpenShadowRoot is added, UserAgentShadowRoot should exists as the oldest Shadow Root.</p>
<div id='container'>
    <form>
        <select id='select'><option>hoge</option></select>
        <output id='output'></output>
        <keygen id='keygen'></keygen>
    </form>
</div>
<pre id='console'></pre>

<script>
var elems = [
    "select", "keygen"
];

var shadowRoot;
var oldestShadowRoot;
var youngerShadowRoot;
function test()
{
    if (window.internals)
        window.internals.settings.setAuthorShadowDOMForAnyElementEnabled(true);

    for (var i = 0; i < elems.length; ++i) {
        var host = document.getElementById(elems[i]);
        shadowRoot = host.createShadowRoot();
        shadowRoot.innerHTML = "<shadow></shadow>";

        oldestShadowRoot = internals.oldestShadowRoot(host);
        youngerShadowRoot = internals.youngerShadowRoot(oldestShadowRoot);
        shouldNotBe('oldestShadowRoot', 'shadowRoot');
        shouldBe('youngerShadowRoot', 'shadowRoot');
    }

    container.innerHTML = "";
}

test();
var successfullyParsed = true;
</script>
</body>
</html>
